fix: update VIP-390 with final changes
